# admin-app/public/openapi.yaml
openapi: 3.0.3
info:
  title: Broker AI API
  version: 1.0.0
servers:
  - url: https://easymo.vercel.app
    description: Vercel proxy with path prefixes
paths:
...
  /api/wa/outbound/messages:
    post:
      summary: Send WhatsApp text or template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to: { type: string }
                text: { type: string }
                template: { type: object }
      responses:
        '202': { description: Accepted }
...